basic rhodes update-1 ci checks
