+++ tests/network-tests/src/misc/postRuntimUpdateChecks.ts
@@ -1,0 +1,14 @@
+import { assert } from 'chai'
+import { FlowProps } from '../Flow'
+import { extendDebug } from '../Debugger'
+
+export default async function assertValues({ api }: FlowProps): Promise<void> {
+  const debug = extendDebug('flow:postMigrationAssertions')
+  debug('Started')
+
+  debug('Check runtime spec version')
+  const version = await api.rpc.state.getRuntimeVersion()
+  assert.equal(version.specVersion.toNumber(), 7)
+
+  debug('Done')
+}
