+++ tests/network-tests/src/scenarios/postRuntimeUpdate.ts
@@ -1,7 +1,7 @@
 import { scenario } from '../Scenario'
-import assertValues from '../misc/assertPostUpgradeConsts'
+import postRuntimeUpdateChecks from '../misc/postRuntimUpdateChecks'
 
 scenario('Post Runtime Upgrade', async ({ job }) => {
   // Verify constants
-  job('Verify Values', assertValues)
+  job('Run Checks', postRuntimeUpdateChecks)
 })
